moved tests into seperate folder
